requirements.txt
# requirements.txt
streamlit
python-dotenv
transformers
sentence-transformers
PyPDF2
python-docx
pandas
plotly-express